test: moved throttle tests to suite
